+++ Modules/Training/Providers/TrainingServiceProvider.php
@@ -8,6 +8,7 @@ use Modules\Training\Livewire\TrainingList;
 use Modules\Training\Livewire\TrainingCalendar;
 use Modules\Training\Livewire\RecurringTrainings;
 use Modules\Training\Livewire\VenueList;
+use Modules\Training\Livewire\AnnouncementCreate;
 
 class TrainingServiceProvider extends ServiceProvider
 {
@@ -26,6 +27,7 @@ class TrainingServiceProvider extends ServiceProvider
         Livewire::component('training.training-calendar', TrainingCalendar::class);
         Livewire::component('training.recurring-trainings', RecurringTrainings::class);
         Livewire::component('training.venue-list', VenueList::class);
+        Livewire::component('training.announcement-create', AnnouncementCreate::class);
     }
 
     public function register(): void